refactor: split the five generators and the split planner along their seams
boxBuilder, labelTabBuilder, compartmentBuilder, kumikoWrapBuilder,
assemblyGenerator and splitPlanner each hand a self-contained group of helpers
to sibling modules: top shape; divider clips and socket cuts; cavities and
divider segments; perimeter, segments and the two slab cutters; part template
and part label; tiling and margins. Public names re-export from their
original modules.
