+++ src/features/generation/worker/generators/compartmentBuilder.ts
@@ -5,21 +5,9 @@
  * Walls appear at boundaries between cells with different compartment IDs.
  */
 
-import {
-  box,
-  withScope,
-  clone,
-  unwrap,
-  fuseAll,
-  draw,
-  intersect,
-  cut,
-  rotate,
-  translate,
-  applyMatrix,
-} from 'brepjs';
-import type { Shape3D, ValidSolid, DisposalScope, Drawing } from 'brepjs';
-import type { BinParams, DividerOverride } from '@/shared/types/bin';
+import { box, withScope, clone, unwrap, fuseAll, cut } from 'brepjs';
+import type { Shape3D, ValidSolid, DisposalScope } from 'brepjs';
+import type { BinParams } from '@/shared/types/bin';
 // Pure grid helpers, kept in the compartment-grid util so the main thread can
 // reach them too — the lid's click rails notch around the same runs
 // and cannot import this module, which pulls in brepjs.
@@ -32,8 +20,6 @@ import {
 
 export { buildOverrideLookup, findPairAwareRuns, overrideKey };
 import { buildCacheKey, compactKey, quantize, stableSerialize } from './cacheKeyUtils';
-import { sketch } from './meshUtils';
-import { BOX_CORNER_RADIUS } from './generatorConstants';
 import { resolveCompartmentDividerHeight } from '@/shared/utils/slotMath';
 import { isAbortError } from './utils/abort';
 import {
@@ -45,6 +31,16 @@ import type { SpanningDividerClip } from './labelTabBuilder';
 
 // Re-export for backwards compatibility with existing imports
 export { fuseAllOrNull } from './utils/shapeOps';
+import { findCompartmentBounds } from './compartmentCavities';
+export {
+  buildCompartmentCavityDrawings,
+  findCompartmentBounds,
+  compartmentCornersRoundCleanly,
+  compartmentCavitiesAreViableWithOverrides,
+} from './compartmentCavities';
+import { buildTiltedWallSegment } from './compartmentDividerSegments';
+export { interiorDividerSegments, findWallSegments } from './compartmentDividerSegments';
+export type { InteriorDividerSegment } from './compartmentDividerSegments';
 
 /**
  * Whether the bin's compartments produce internal divider walls (>1 distinct
@@ -55,184 +51,6 @@ export function hasMultipleCompartments(params: BinParams): boolean {
   const { cells } = params.compartments;
   if (cells.length === 0) return false;
   return new Set(cells).size > 1;
-}
-
-/**
- * Build per-compartment cavity drawings (rectangular footprints in the XY
- * plane) used to subtract cavities from the bin's outer extrusion, producing
- * the divider walls as natural cut residue between compartments.
- *
- * Each cavity rectangle spans the compartment's cell bounding box, inset by
- * `thickness/2` on every shared internal boundary (so adjacent compartments
- * leave `thickness` of wall between them) and aligned flush with the bin's
- * inner wall surface on each exterior boundary.
- *
- * Corners that sit on the bin perimeter are rounded with the same radius the
- * single-compartment hollow shell uses (`BOX_CORNER_RADIUS − wallThickness`)
- * so the cavity follows the rounded inner wall contour. Without this, a sharp
- * cavity corner pokes past the outer rounded arc on thin-walled bins
- * (`wallThickness < BOX_CORNER_RADIUS·(1 − 1/√2) ≈ 1.1mm`) and the cut eats
- * through the outer skin, leaving a gap in the bin's corners.
- *
- * Non-rectangular compartments (cells with the same ID forming an L-shape,
- * etc.) are approximated by their bounding box; multi-cavity cut is therefore
- * only safe for compartments that fill their bounding box. Callers should
- * verify via `compartmentsAreRectangular` before using this path.
- */
-export function buildCompartmentCavityDrawings(
-  params: BinParams,
-  innerW: number,
-  innerD: number
-): readonly Drawing[] {
-  const lookup = buildOverrideLookup(params.compartments.dividerOverrides);
-  const cornerRadius = Math.max(BOX_CORNER_RADIUS - params.wallThickness, 0);
-  const drawings: Drawing[] = [];
-  for (const id of new Set(params.compartments.cells)) {
-    const corners = cavityCorners(params, innerW, innerD, id, lookup);
-    if (!corners) continue;
-    drawings.push(cavityDrawing(corners, cornerRadius));
-  }
-  return drawings;
-}
-
-/**
- * Build a cavity drawing, rounding each corner that lies on the bin
- * perimeter with `cornerRadius` (clamped to the cavity's own dimensions so
- * the fillet always fits). Interior corners — divider junctions — stay sharp.
- */
-function cavityDrawing(corners: CavityCorners, cornerRadius: number): Drawing {
-  const { bl, br, tr, tl, exterior } = corners;
-  // Cap the radius at half the smaller span so opposing fillets can't overrun.
-  const { cavW, cavD } = cavitySpan(corners);
-  const r = Math.max(0, Math.min(cornerRadius, cavW / 2 - 0.05, cavD / 2 - 0.05));
-  const rBL = exterior.bl ? r : 0;
-  const rBR = exterior.br ? r : 0;
-  const rTR = exterior.tr ? r : 0;
-  const rTL = exterior.tl ? r : 0;
-
-  if (r <= 0.1 || !hasExteriorCorner(exterior)) {
-    return draw(bl).lineTo(br).lineTo(tr).lineTo(tl).close();
-  }
-
-  // Start at the midpoint of BL→BR so close() forms a real edge through BL,
-  // letting customCorner(rBL) apply to it (mirrors buildSlabProfile in
-  // baseplateSlab.ts). Starting at a corner would make close() degenerate.
-  // Use the true midpoint (not [mid_x, bl[1]]): an override can tilt the
-  // bottom edge (bl[1] !== br[1]), and the start point must stay on it.
-  let pen = draw([(bl[0] + br[0]) / 2, (bl[1] + br[1]) / 2]);
-  pen = pen.lineTo(br);
-  if (rBR > 0) pen = pen.customCorner(rBR);
-  pen = pen.lineTo(tr);
-  if (rTR > 0) pen = pen.customCorner(rTR);
-  pen = pen.lineTo(tl);
-  if (rTL > 0) pen = pen.customCorner(rTL);
-  pen = pen.lineTo(bl);
-  if (rBL > 0) pen = pen.customCorner(rBL);
-  return pen.close();
-}
-
-interface CavityCorners {
-  bl: [number, number];
-  br: [number, number];
-  tr: [number, number];
-  tl: [number, number];
-  /**
-   * Whether each corner sits on the bin perimeter (both adjacent edges are
-   * exterior). Only these corners get rounded; interior corners are divider
-   * junctions and stay sharp.
-   */
-  exterior: { bl: boolean; br: boolean; tr: boolean; tl: boolean };
-}
-
-/** Whether any cavity corner sits on the bin perimeter (and so gets rounded). */
-function hasExteriorCorner(exterior: CavityCorners['exterior']): boolean {
-  return exterior.bl || exterior.br || exterior.tr || exterior.tl;
-}
-
-/**
- * Width and depth of a cavity, taken as the smaller of each pair of opposing
- * edges so override displacement can't inflate the span used for fillet
- * clamping or clearance checks.
- */
-function cavitySpan(corners: CavityCorners): { cavW: number; cavD: number } {
-  const { bl, br, tr, tl } = corners;
-  return {
-    cavW: Math.min(br[0] - bl[0], tr[0] - tl[0]),
-    cavD: Math.min(tl[1] - bl[1], tr[1] - br[1]),
-  };
-}
-
-/**
- * Cavity corners with shared-edge overrides applied. Sign convention
- * mirrors `buildTiltedWallSegment`: `offsetStart` shifts the lower-
- * coordinate endpoint, `offsetEnd` the higher-coordinate one. Both
- * compartments either side of a divider apply the same offsets so the
- * centerline displaces consistently and each side keeps its `half` inset.
- */
-function cavityCorners(
-  params: BinParams,
-  innerW: number,
-  innerD: number,
-  id: number,
-  lookup: Map<string, DividerOverride>
-): CavityCorners | null {
-  const { cols, rows, thickness, cells } = params.compartments;
-  const bounds = findCompartmentBounds(id, cols, rows, cells);
-  if (!bounds) return null;
-  const { minCol, maxCol, minRow, maxRow } = bounds;
-  const cellW = innerW / cols;
-  const cellD = innerD / rows;
-  const half = thickness / 2;
-  const xMin = -innerW / 2 + minCol * cellW + (minCol > 0 ? half : 0);
-  const xMax = -innerW / 2 + (maxCol + 1) * cellW - (maxCol < cols - 1 ? half : 0);
-  const yMin = -innerD / 2 + minRow * cellD + (minRow > 0 ? half : 0);
-  const yMax = -innerD / 2 + (maxRow + 1) * cellD - (maxRow < rows - 1 ? half : 0);
-  const bl: [number, number] = [xMin, yMin];
-  const br: [number, number] = [xMax, yMin];
-  const tr: [number, number] = [xMax, yMax];
-  const tl: [number, number] = [xMin, yMax];
-  // A corner is on the bin perimeter when both of its edges are exterior.
-  // Override displacement only touches interior edges, so exterior corners
-  // are never shifted (safe to round).
-  const left = minCol === 0;
-  const right = maxCol === cols - 1;
-  const front = minRow === 0;
-  const back = maxRow === rows - 1;
-  const exterior = {
-    bl: left && front,
-    br: right && front,
-    tr: right && back,
-    tl: left && back,
-  };
-  if (maxCol < cols - 1) {
-    const ov = lookup.get(overrideKey(id, cells[minRow * cols + (maxCol + 1)]));
-    if (ov) {
-      br[0] += ov.offsetStart;
-      tr[0] += ov.offsetEnd;
-    }
-  }
-  if (minCol > 0) {
-    const ov = lookup.get(overrideKey(id, cells[minRow * cols + (minCol - 1)]));
-    if (ov) {
-      bl[0] += ov.offsetStart;
-      tl[0] += ov.offsetEnd;
-    }
-  }
-  if (maxRow < rows - 1) {
-    const ov = lookup.get(overrideKey(id, cells[(maxRow + 1) * cols + minCol]));
-    if (ov) {
-      tl[1] += ov.offsetStart;
-      tr[1] += ov.offsetEnd;
-    }
-  }
-  if (minRow > 0) {
-    const ov = lookup.get(overrideKey(id, cells[(minRow - 1) * cols + minCol]));
-    if (ov) {
-      bl[1] += ov.offsetStart;
-      br[1] += ov.offsetEnd;
-    }
-  }
-  return { bl, br, tr, tl, exterior };
 }
 
 /** Whether any divider override is present in the compartment grid. */
@@ -295,33 +113,6 @@ export function compartmentEdgesAreSinglePair(params: BinParams): boolean {
 }
 
 /**
- * Verify every override-displaced cavity remains a non-degenerate quad
- * (left strictly left of right, bottom strictly below top, with at least
- * `thickness * 2` clearance). Extreme offsets can otherwise produce a
- * self-intersecting bowtie that BREP silently drops; falling back to the
- * additive-fuse path lets that path's clip-to-interior salvage *some*
- * mesh for pathological inputs.
- */
-export function compartmentCavitiesAreViableWithOverrides(
-  params: BinParams,
-  innerW: number,
-  innerD: number
-): boolean {
-  const overrides = params.compartments.dividerOverrides;
-  if (!overrides || overrides.length === 0) return true;
-  const lookup = buildOverrideLookup(overrides);
-  const minDim = params.compartments.thickness * 2;
-  for (const id of new Set(params.compartments.cells)) {
-    const corners = cavityCorners(params, innerW, innerD, id, lookup);
-    if (!corners) continue;
-    const { bl, br, tr, tl } = corners;
-    if (Math.min(br[0], tr[0]) - Math.max(bl[0], tl[0]) < minDim) return false;
-    if (Math.min(tl[1], tr[1]) - Math.max(bl[1], br[1]) < minDim) return false;
-  }
-  return true;
-}
-
-/**
  * Whether every compartment is a rectangle (its cells fully fill its
  * bounding box). Required precondition for the multi-cavity-cut shell
  * path; non-rectangular compartments fall back to the additive-fuse path.
@@ -378,370 +169,11 @@ export function compartmentCavitiesAreViable(
   return true;
 }
 
-/**
- * Whether every bin-perimeter cavity corner can be rounded with the full
- * inner-shell radius (`BOX_CORNER_RADIUS − wallThickness`). On thin walls a
- * corner compartment narrower than twice that radius can only be partially
- * rounded, so its sharp residue still pokes past the outer arc and reopens
- * the corner gap — such bins must fall back to the additive-fuse path
- * (whose rounded hollow shell has no corner gap). Thick-walled bins are
- * always safe: a sharp cavity corner never reaches past the arc, so cell
- * size is irrelevant.
- */
-export function compartmentCornersRoundCleanly(
-  params: BinParams,
-  innerW: number,
-  innerD: number
-): boolean {
-  const targetR = BOX_CORNER_RADIUS - params.wallThickness;
-  // Sharp corners only over-cut below BOX_CORNER_RADIUS·(1 − 1/√2); above
-  // that threshold there is no gap to fix regardless of compartment size.
-  if (targetR <= BOX_CORNER_RADIUS / Math.SQRT2) return true;
-  const need = 2 * targetR;
-  const lookup = buildOverrideLookup(params.compartments.dividerOverrides);
-  for (const id of new Set(params.compartments.cells)) {
-    const corners = cavityCorners(params, innerW, innerD, id, lookup);
-    if (!corners) continue;
-    if (!hasExteriorCorner(corners.exterior)) continue;
-    const { cavW, cavD } = cavitySpan(corners);
-    if (cavW < need || cavD < need) return false;
-  }
-  return true;
-}
-
 /** Build a positioned wall segment solid. */
 function buildWallSegment(w: number, d: number, height: number, x: number, y: number): Shape3D {
   return box(w, d, height, { at: [x, y, height / 2] });
 }
 
-/** One divider wall segment: its top line, how far its foot leans off that
- *  line, and the interior it gets clipped to. */
-interface WallSegmentPlan {
-  readonly startX: number;
-  readonly startY: number;
-  readonly endX: number;
-  readonly endY: number;
-  /** Foot displacement from the top line, along the offset axis. Zero stands
-   *  the wall upright. */
-  readonly driftX: number;
-  readonly driftY: number;
-  readonly thickness: number;
-  readonly height: number;
-  readonly binInnerW: number;
-  readonly binInnerD: number;
-}
-
-/**
- * Build a tilted divider wall whose top edge runs from `(startX, startY)` to
- * `(endX, endY)` and whose foot edge is that top edge translated by the plan's
- * full drift vector. Thickness is applied perpendicular to the wall PLANE (not
- * world-aligned) so the divider reads as a tilted ribbon of constant thickness,
- * not a squished box — true even when the wall is both angled in plan and leaned
- * off vertical (a general oblique parallelepiped, not just a parallelogram
- * prism).
- *
- * Clipped to the bin interior so any corners that overshoot the bin wall (which
- * happens whenever the tilt is non-zero) are sliced cleanly at the wall plane.
- */
-function buildTiltedWallSegment(scope: DisposalScope, plan: WallSegmentPlan): Shape3D | null {
-  const { startX, startY, endX, endY, thickness, height, binInnerW, binInnerD } = plan;
-  const dx = endX - startX;
-  const dy = endY - startY;
-  const len = Math.hypot(dx, dy);
-  // Degenerate-segment guard: a zero-length divider (both endpoints
-  // collapsed to the same point — possible if a malformed override drives
-  // a span to nothing) would divide by zero and produce NaN geometry.
-  if (len < 1e-6) return null;
-  // Perpendicular unit vector (rotated 90° CCW from the divider direction).
-  const px = -dy / len;
-  const py = dx / len;
-  const half = thickness / 2;
-
-  // Split the foot's drift into components across and along the run. The foot
-  // line is the top line translated by the full drift vector, so BOTH matter:
-  // `across` leans the wall off vertical; `along` shears it lengthwise (the
-  // foot line slides parallel to itself down the run). When a divider is only
-  // angled OR only leaned `along` is zero, but combine Angle and Lean and the
-  // offset axis no longer lines up with the run, so a real along-run component
-  // appears. Dropping it (keeping only `across`) is what put a compound
-  // divider's foot outside the bin: the wall then leaned the wrong way and the
-  // interior clip sliced it into a wedge.
-  const leanAcross = plan.driftX * px + plan.driftY * py;
-  const driftAlong = (plan.driftX * dx + plan.driftY * dy) / len;
-
-  let prism: Shape3D;
-  if (leanAcross === 0 && driftAlong === 0) {
-    const pen = draw([startX + px * half, startY + py * half])
-      .lineTo([endX + px * half, endY + py * half])
-      .lineTo([endX - px * half, endY - py * half])
-      .lineTo([startX - px * half, startY - py * half])
-      .close();
-    prism = scope.register(sketch(pen, 'XY', 0).extrude(height));
-  } else {
-    // `buildLeaningPrism` builds the across-run lean in a canonical frame (run
-    // along +X, top edge on the X axis at `height`), preserving perpendicular
-    // thickness. The along-run drift is then a shear parallel to the run,
-    // `x += driftAlong · (1 − z/height)`: the top edge (z=height) stays put and
-    // the foot (z=0) slides `driftAlong` down +X. Because the shear direction
-    // lies in the wall plane, it does not change the plane's normal and so keeps
-    // the perpendicular thickness exactly `thickness` (see __kernel-tests__/
-    // dividerRake). Composing lean + shear reproduces the true foot line where a
-    // single rigid tilt cannot.
-    const canonical = scope.register(buildLeaningPrism(len, leanAcross, thickness, height));
-    const sheared =
-      driftAlong === 0
-        ? canonical
-        : scope.register(
-            unwrap(
-              applyMatrix(canonical, [
-                [1, 0, -driftAlong / height, driftAlong],
-                [0, 1, 0, 0],
-                [0, 0, 1, 0],
-                [0, 0, 0, 1],
-              ])
-            )
-          );
-    const oriented = scope.register(
-      rotate(sheared, (Math.atan2(dy, dx) * 180) / Math.PI, { axis: [0, 0, 1] })
-    );
-    prism = scope.register(translate(oriented, [(startX + endX) / 2, (startY + endY) / 2, 0]));
-  }
-  // Clip the parallelogram corners that overshoot the bin's perpendicular
-  // wall. For zero-tilt segments the clip is a no-op; for tilted segments
-  // it shears off the "ears" that would otherwise poke through the wall.
-  // If the prism is entirely outside the clip box (override pushes the
-  // divider beyond the bin interior), the intersect returns nothing — bail
-  // gracefully rather than crashing the whole bin build.
-  const clipBox = scope.register(box(binInnerW, binInnerD, height, { at: [0, 0, height / 2] }));
-  try {
-    return scope.register(unwrap(intersect(prism as ValidSolid, clipBox)));
-  } catch {
-    return null;
-  }
-}
-
-/**
- * A leaning wall in canonical pose: run along +X centred on the origin, top
- * edge on the X axis at `height`, foot displaced `leanAcross` in +Y.
- *
- * The section is a parallelogram with flat ends at z=0 and z=height and
- * horizontal half-width `t/2 · L/height`, which is what keeps the distance
- * between its two slanted faces at exactly `thickness` however far it leans.
- * Sizing the section on the plan footprint instead thins the wall by
- * cos(lean), taking a 1.6mm divider under two perimeters at 45°.
- */
-function buildLeaningPrism(
-  len: number,
-  leanAcross: number,
-  thickness: number,
-  height: number
-): Shape3D {
-  const half = (thickness / 2) * (Math.hypot(leanAcross, height) / height);
-  // 2D sketch axes on YZ are (world Y, world Z); the extrude runs along +X.
-  const section = draw([-half, height])
-    .lineTo([half, height])
-    .lineTo([leanAcross + half, 0])
-    .lineTo([leanAcross - half, 0])
-    .close();
-  return section.sketchOnPlane('YZ', -len / 2).extrude(len);
-}
-
-/** One interior divider wall segment, resolved to mm in bin-centered coords. */
-export interface InteriorDividerSegment {
-  /** Axis-projected segment length (mm) — not the longer true diagonal. */
-  readonly segLen: number;
-  /** True wall length along the (possibly tilted) segment — equals `segLen`
-   *  for straight dividers, longer for tilted ones. Features that measure
-   *  distance ALONG the wall (cutout width, alignment) must use this, not the
-   *  projected `segLen`. */
-  readonly wallLen: number;
-  readonly x: number;
-  readonly y: number;
-  /** In-plane rotation (deg) aligned to the wall: 90 for straight vertical
-   *  dividers, 0 for straight horizontal, tilted by the override otherwise. */
-  readonly rotateZ: number;
-  /** Lean off vertical (deg). `x`/`y` describe the wall's TOP edge, so any
-   *  feature that reaches below the rim has to answer for this: on a leaning
-   *  divider the wall is simply not at `x`/`y` further down. */
-  readonly leanDeg: number;
-  /** Where the wall meets the FLOOR. Equals `x`/`y` when it stands upright. */
-  readonly footX: number;
-  readonly footY: number;
-}
-
-/**
- * Enumerate every interior divider wall segment with its tilt-resolved
- * placement, honouring `dividerOverrides`. Shared source of truth so features
- * that decorate dividers (wall cutouts, handles) land ON the wall instead of at
- * the original grid line when a divider is tilted. Pure — no WASM.
- */
-export function interiorDividerSegments(
-  params: BinParams,
-  innerW: number,
-  innerD: number,
-  dividerHeight: number
-): InteriorDividerSegment[] {
-  const { cols, rows, cells } = params.compartments;
-  const out: InteriorDividerSegment[] = [];
-  if (cols <= 1 && rows <= 1) return out;
-
-  const cellW = innerW / cols;
-  const cellD = innerD / rows;
-  const lookup = buildOverrideLookup(params.compartments.dividerOverrides);
-  const RAD2DEG = 180 / Math.PI;
-
-  // Without overrides, merge contiguous wall cells into one run (historical
-  // behavior — one window per span). Only split per compartment pair when tilts
-  // exist, since each tilted pair is its own angled wall segment.
-  const hasOverrides = lookup.size > 0;
-  const runsFor = (
-    count: number,
-    pairOf: (i: number) => string | null
-  ): Array<{ start: number; end: number; pairKey: string }> =>
-    hasOverrides
-      ? findPairAwareRuns(count, pairOf)
-      : findWallSegments(count, (i) => pairOf(i) !== null).map(([start, end]) => ({
-          start,
-          end,
-          pairKey: '',
-        }));
-
-  // Vertical dividers (between columns) run along Y; straight ⇒ rotateZ 90.
-  for (let boundary = 1; boundary < cols; boundary++) {
-    const xPos = -innerW / 2 + boundary * cellW;
-    const runs = runsFor(rows, (row) => {
-      const leftId = cells[row * cols + (boundary - 1)];
-      const rightId = cells[row * cols + boundary];
-      return leftId !== rightId ? overrideKey(leftId, rightId) : null;
-    });
-    for (const { start, end, pairKey } of runs) {
-      const segLen = (end - start) * cellD;
-      const midY = -innerD / 2 + (start + (end - start) / 2) * cellD;
-      const ov = lookup.get(pairKey);
-      out.push(
-        ov
-          ? (() => {
-              const x = xPos + (ov.offsetStart + ov.offsetEnd) / 2;
-              return {
-                segLen,
-                wallLen: Math.hypot(segLen, ov.offsetEnd - ov.offsetStart),
-                x,
-                y: midY,
-                rotateZ: Math.atan2(segLen, ov.offsetEnd - ov.offsetStart) * RAD2DEG,
-                leanDeg: ov.rakeDeg ?? 0,
-                footX: x + dividerFootDrift(ov, dividerHeight),
-                footY: midY,
-              };
-            })()
-          : {
-              segLen,
-              wallLen: segLen,
-              x: xPos,
-              y: midY,
-              rotateZ: 90,
-              leanDeg: 0,
-              footX: xPos,
-              footY: midY,
-            }
-      );
-    }
-  }
-
-  // Horizontal dividers (between rows) run along X; straight ⇒ rotateZ 0.
-  for (let boundary = 1; boundary < rows; boundary++) {
-    const yPos = -innerD / 2 + boundary * cellD;
-    const runs = runsFor(cols, (col) => {
-      const topId = cells[(boundary - 1) * cols + col];
-      const bottomId = cells[boundary * cols + col];
-      return topId !== bottomId ? overrideKey(topId, bottomId) : null;
-    });
-    for (const { start, end, pairKey } of runs) {
-      const segLen = (end - start) * cellW;
-      const midX = -innerW / 2 + (start + (end - start) / 2) * cellW;
-      const ov = lookup.get(pairKey);
-      out.push(
-        ov
-          ? (() => {
-              const y = yPos + (ov.offsetStart + ov.offsetEnd) / 2;
-              return {
-                segLen,
-                wallLen: Math.hypot(segLen, ov.offsetEnd - ov.offsetStart),
-                x: midX,
-                y,
-                rotateZ: Math.atan2(ov.offsetEnd - ov.offsetStart, segLen) * RAD2DEG,
-                leanDeg: ov.rakeDeg ?? 0,
-                footX: midX,
-                footY: y + dividerFootDrift(ov, dividerHeight),
-              };
-            })()
-          : {
-              segLen,
-              wallLen: segLen,
-              x: midX,
-              y: yPos,
-              rotateZ: 0,
-              leanDeg: 0,
-              footX: midX,
-              footY: yPos,
-            }
-      );
-    }
-  }
-  return out;
-}
-
-/**
- * Find consecutive wall segments along a boundary line.
- * Returns array of [start, end) index pairs where walls are needed.
- */
-export function findWallSegments(
-  count: number,
-  needsWall: (i: number) => boolean
-): Array<[number, number]> {
-  const segments: Array<[number, number]> = [];
-  let segStart: number | null = null;
-
-  for (let i = 0; i < count; i++) {
-    if (needsWall(i)) {
-      if (segStart === null) segStart = i;
-    } else if (segStart !== null) {
-      segments.push([segStart, i]);
-      segStart = null;
-    }
-  }
-  if (segStart !== null) {
-    segments.push([segStart, count]);
-  }
-  return segments;
-}
-
-/**
- * Find the bounding row/column range of a compartment by its ID.
- * Returns null if the compartment ID is not found in the grid.
- */
-export function findCompartmentBounds(
-  compId: number,
-  cols: number,
-  rows: number,
-  cells: readonly number[]
-): { minCol: number; maxCol: number; minRow: number; maxRow: number } | null {
-  let minCol = cols;
-  let maxCol = -1;
-  let minRow = rows;
-  let maxRow = -1;
-  for (let r = 0; r < rows; r++) {
-    for (let c = 0; c < cols; c++) {
-      if (cells[r * cols + c] === compId) {
-        minCol = Math.min(minCol, c);
-        maxCol = Math.max(maxCol, c);
-        minRow = Math.min(minRow, r);
-        maxRow = Math.max(maxRow, r);
-      }
-    }
-  }
-  if (maxCol === -1) return null;
-  return { minCol, maxCol, minRow, maxRow };
-}
 /**
  * Build compartment divider walls inside the bin.
  *
