+++ src/features/generation/worker/generators/labelTabBuilder.ts
@@ -7,7 +7,6 @@
 
 import { labelShelfKeepoutMm } from '@/shared/utils/lidInteriorRelief';
 import {
-  box,
   draw,
   drawRoundedRectangle,
   unwrap,
@@ -20,27 +19,19 @@ import {
   clone,
 } from 'brepjs';
 import type { Shape3D, ValidSolid, Drawing, DisposalScope } from 'brepjs';
-import { BOX_CORNER_RADIUS, COPLANAR_MARGIN, COPLANAR_OVERLAP } from './generatorConstants';
+import { BOX_CORNER_RADIUS, COPLANAR_OVERLAP } from './generatorConstants';
 import type { BinParams, TextStyleDefaults, TextStyleOverride } from '@/shared/types/bin';
 import { resolveTextStyle } from '@/shared/types/bin';
 import {
-  LABEL_PLATE_CORNER_RADIUS_MM,
   LABEL_PLATE_HEIGHT_MM,
   LABEL_SOCKET_CLICK_POCKET_DEPTH_MM,
-  LABEL_SOCKET_DETENT_DEPTH_MM,
-  LABEL_SOCKET_DETENT_HEIGHT_MM,
-  LABEL_SOCKET_LIP_OVERHANG_MM,
-  LABEL_SOCKET_LIP_THICKNESS_MM,
   LABEL_SOCKET_POCKET_DEPTH_MM,
-  LABEL_SOCKET_RIB_HEIGHT_MM,
-  LABEL_SOCKET_RIB_PROTRUSION_MM,
-  LABEL_SOCKET_RIB_START_MM,
   LABEL_SOCKET_SLIDE_Z_CLEARANCE_MM,
   LABEL_SOCKET_WALL_MM,
   labelLipReservationMm,
   labelPlateWidthMm,
 } from '@/shared/constants/labelPlates';
-import type { LabelPlateWidthU, LabelSocketStyle } from '@/shared/constants/labelPlates';
+import type { LabelPlateWidthU } from '@/shared/constants/labelPlates';
 import { NOZZLE_BASELINE } from '@/shared/printSettings/connectorScaling';
 import { planLabelTabLayout } from '@/shared/utils/labelTabPlan';
 import type { TabSlot, PlannedTabRow, TabBuildDimensions } from '@/shared/utils/labelTabPlan';
@@ -50,6 +41,14 @@ import { sketch } from './meshUtils';
 import { buildFilletProfile } from './filletProfile';
 import { buildTextSolid, fitTextSize } from './textBuilder';
 import type { LabelTextOverflow } from '../../bridge/types';
+export {
+  planSpanningDividerClips,
+  buildSpanningDividerClipTools,
+  spanningDividerClipsKey,
+} from './spanningDividerClips';
+export type { SpanningDividerClip } from './spanningDividerClips';
+import { applySocket } from './labelSocketCutter';
+export { cutLabelSocket } from './labelSocketCutter';
 
 /** Tab text fills its shelf band rather than the font's line box. Shared by the
  *  group size pass and the per-tab build — measuring different boxes would
@@ -110,102 +109,6 @@ export function buildLabelTabs(
     const fused = buildLabelTabsInScope(scope, params, innerW, innerD, wallHeight, wallThickness);
     return fused ? unwrap(clone(fused)) : null;
   });
-}
-
-/**
- * A footprint a wall-to-wall shelf passes over, in the bin-interior frame.
- * `zMin` is the shelf underside: divider material from there up is what the
- * shelf would otherwise collide with.
- */
-export interface SpanningDividerClip {
-  readonly xMin: number;
-  readonly xMax: number;
-  readonly yMin: number;
-  readonly yMax: number;
-  readonly zMin: number;
-}
-
-/**
- * Where a full-width shelf crosses the column dividers.
- *
- * `planSpanningTabAtRow` already assumes those dividers "pass beneath" the
- * span — but nothing ever shortened them, so they ran to the interior ceiling
- * and stood proud of any shelf sunk below it (a click-in socket's stacking
- * relief, or an explicit `label.height`), splitting the one continuous label
- * surface the feature exists to provide.
- *
- * Derived from the same layout plan the shelves themselves are built from, so
- * the clip and the shelf cannot drift apart. Both spanning shapes qualify: the
- * `label.span` feature and the socket plan's bin-spanning fallback.
- */
-export function planSpanningDividerClips(
-  params: BinParams,
-  innerW: number,
-  innerD: number,
-  wallHeight: number,
-  wallThickness: number
-): SpanningDividerClip[] {
-  if (!params.label.enabled) return [];
-  const layout = planLabelTabLayout(params, innerW, innerD, wallHeight, wallThickness);
-  if (!layout) return [];
-  // Per-compartment tabs are bounded by the dividers rather than crossing
-  // them, so there is nothing to clip.
-  if (!layout.spanningFallback && params.label.span !== true) return [];
-
-  const { dims } = layout;
-  const zMin = dims.shelfTopZ - dims.shelfT;
-  const clips: SpanningDividerClip[] = [];
-  for (const row of layout.plannedRows) {
-    const depthSign = row.anchor === 'back' ? -1 : 1;
-    for (const slot of row.slots) {
-      const yEnd = slot.positionY + depthSign * dims.tabDepth;
-      clips.push({
-        xMin: slot.tabXStart,
-        xMax: slot.tabXStart + slot.tabWidth,
-        yMin: Math.min(slot.positionY, yEnd),
-        yMax: Math.max(slot.positionY, yEnd),
-        zMin,
-      });
-    }
-  }
-  return clips;
-}
-
-/**
- * Cut tools for a clip set, each running from the shelf underside to `topZ`.
- *
- * `topZ` must clear whatever the caller is cutting (the interior ceiling, or a
- * collared rim) — the box only has to swallow the divider top, and overshooting
- * upward hits empty space.
- */
-export function buildSpanningDividerClipTools(
-  clips: readonly SpanningDividerClip[],
-  topZ: number,
-  offsetX = 0,
-  offsetY = 0
-): Shape3D[] {
-  const tools: Shape3D[] = [];
-  for (const c of clips) {
-    const height = topZ - c.zMin;
-    if (height <= 0) continue;
-    const w = c.xMax - c.xMin;
-    const d = c.yMax - c.yMin;
-    if (w <= 0 || d <= 0) continue;
-    tools.push(
-      box(w, d, height, {
-        at: [(c.xMin + c.xMax) / 2 + offsetX, (c.yMin + c.yMax) / 2 + offsetY, c.zMin + height / 2],
-      })
-    );
-  }
-  return tools;
-}
-
-/** Cache-key segment for a clip set. Empty string when nothing is clipped. */
-export function spanningDividerClipsKey(clips: readonly SpanningDividerClip[]): string {
-  if (clips.length === 0) return '';
-  return clips
-    .map((c) => [c.xMin, c.xMax, c.yMin, c.yMax, c.zMin].map((n) => n.toFixed(3)).join(','))
-    .join(';');
 }
 
 /**
@@ -687,233 +590,6 @@ function buildTabsAtRow(
     result.push(tabSolid);
   }
 
-  return result;
-}
-
-/**
- * Cut a swappable-label socket into the shelf top and fuse the retention
- * ribs. Local tab frame: shelf spans X:[0,tabWidth],
- * Y:[depthSign·tabDepth, 0] with the shelf top at Z=tabHeight.
- *
- * Pocket = plate footprint + total clearance, one pocket-wall margin in
- * from the anchor wall, placed along X by `alignment`. Ribs sit on the two
- * long (X-parallel) pocket walls: 0.2mm proud, 0.4mm tall, starting 0.2mm
- * above the pocket floor — the band the plate's perimeter latch clicks
- * behind.
- *
- * Best-effort like `applyTabText`: geometry that doesn't fit or a boolean
- * that throws leaves the plain shelf rather than tanking the tab build.
- */
-function applySocket(
-  scope: DisposalScope,
-  tabSolid: Shape3D,
-  ctx: {
-    plateWidthU: LabelPlateWidthU;
-    clearanceMm: number;
-    style: LabelSocketStyle;
-    tabWidth: number;
-    tabDepth: number;
-    tabHeight: number;
-    alignment: 'left' | 'center' | 'right';
-    depthSign: 1 | -1;
-  }
-): Shape3D {
-  const wall = LABEL_SOCKET_WALL_MM;
-  const pocketW = labelPlateWidthMm(ctx.plateWidthU) + ctx.clearanceMm;
-  const pocketD = LABEL_PLATE_HEIGHT_MM + ctx.clearanceMm;
-
-  // Defense in depth: the plan already sized the plate to the tab width, but
-  // a crafted payload (short depth, huge fit offset) could still overflow.
-  if (pocketW + 2 * wall > ctx.tabWidth + 0.01) return tabSolid;
-  if (pocketD + 2 * wall > ctx.tabDepth + 0.01) return tabSolid;
-
-  let pocketX0: number;
-  if (ctx.alignment === 'left') {
-    pocketX0 = wall;
-  } else if (ctx.alignment === 'right') {
-    pocketX0 = ctx.tabWidth - wall - pocketW;
-  } else {
-    pocketX0 = (ctx.tabWidth - pocketW) / 2;
-  }
-  const centerX = pocketX0 + pocketW / 2;
-  const centerY = ctx.depthSign * (wall + pocketD / 2);
-
-  try {
-    return cutLabelSocket(scope, tabSolid, {
-      centerX,
-      centerY,
-      topZ: ctx.tabHeight,
-      plateWidthU: ctx.plateWidthU,
-      clearanceMm: ctx.clearanceMm,
-      style: ctx.style,
-      // The slide mouth opens through the tab's compartment-facing edge —
-      // extend the cut from the pocket's far edge past the tab boundary.
-      mouth: {
-        sign: ctx.depthSign,
-        extendMm: ctx.tabDepth - wall - pocketD + 1,
-      },
-    });
-  } catch {
-    return tabSolid;
-  }
-}
-
-/**
- * Cut a swappable-label socket into `solid`'s top face, pocket centered at
- * (centerX, centerY). Shared by the label-tab shelf and the fit-calibration
- * coupon so the printed socket can never drift between the two. Throws on
- * boolean failure — callers needing best-effort semantics wrap it.
- *
- * Styles:
- * - `clickIn` (default): pocket + retention ribs, floor at topZ − pocket
- *   depth. The Cullenect-compatible profile.
- * - `slideChannel`: pocket sunk one lip band + z-clearance deeper, with
- *   overhanging lips left/right/anchor-side, a mouth corridor opening
- *   `mouth.sign`-ward through the host's edge (`mouth.extendMm` past the
- *   pocket), and a park detent on the corridor floor at the pocket edge.
- */
-export function cutLabelSocket(
-  scope: DisposalScope,
-  solid: Shape3D,
-  ctx: {
-    centerX: number;
-    centerY: number;
-    topZ: number;
-    plateWidthU: LabelPlateWidthU;
-    clearanceMm: number;
-    style?: LabelSocketStyle;
-    mouth?: { sign: 1 | -1; extendMm: number };
-  }
-): Shape3D {
-  const pocketW = labelPlateWidthMm(ctx.plateWidthU) + ctx.clearanceMm;
-  const pocketD = LABEL_PLATE_HEIGHT_MM + ctx.clearanceMm;
-
-  if ((ctx.style ?? 'clickIn') === 'slideChannel') {
-    if (!ctx.mouth) {
-      // Throwing (not falling back to click-in) keeps a future call site from
-      // silently shipping the wrong retention profile; best-effort callers
-      // already wrap this function.
-      throw new Error('slideChannel socket requires a mouth direction');
-    }
-    return cutSlideChannel(
-      scope,
-      solid,
-      { centerX: ctx.centerX, centerY: ctx.centerY, topZ: ctx.topZ, mouth: ctx.mouth },
-      pocketW,
-      pocketD
-    );
-  }
-
-  const floorZ = ctx.topZ - LABEL_SOCKET_CLICK_POCKET_DEPTH_MM;
-
-  const pocketCutter = scope.register(
-    translate(
-      scope.register(
-        sketch(
-          drawRoundedRectangle(pocketW, pocketD, LABEL_PLATE_CORNER_RADIUS_MM),
-          'XY',
-          floorZ
-        ).extrude(LABEL_SOCKET_CLICK_POCKET_DEPTH_MM + COPLANAR_MARGIN)
-      ),
-      [ctx.centerX, ctx.centerY, 0]
-    )
-  );
-  let result = scope.register(unwrap(cut(solid as ValidSolid, pocketCutter as ValidSolid)));
-
-  // Ribs: full pocket-X span (square ends fuse into the rounded corners,
-  // matching the standard), embedded slightly into the wall so the fuse
-  // never leaves a coplanar seam.
-  const ribEmbed = 0.1;
-  const ribT = LABEL_SOCKET_RIB_PROTRUSION_MM + ribEmbed;
-  const ribZ0 = floorZ + LABEL_SOCKET_RIB_START_MM;
-  const ribProfile = draw([-pocketW / 2, -ribT / 2])
-    .lineTo([pocketW / 2, -ribT / 2])
-    .lineTo([pocketW / 2, ribT / 2])
-    .lineTo([-pocketW / 2, ribT / 2])
-    .close();
-  for (const side of [-1, 1] as const) {
-    const wallY = ctx.centerY + (side * pocketD) / 2;
-    const ribCenterY = wallY - side * (ribT / 2 - ribEmbed);
-    const rib = scope.register(
-      translate(
-        scope.register(sketch(ribProfile, 'XY', ribZ0).extrude(LABEL_SOCKET_RIB_HEIGHT_MM)),
-        [ctx.centerX, ribCenterY, 0]
-      )
-    );
-    result = scope.register(unwrap(fuse(result, rib as ValidSolid)));
-  }
-  return result;
-}
-
-/**
- * Slide-channel variant of `cutLabelSocket`: a two-layer cut (cavity below,
- * lip window above, stacked exactly like the v1 plate channels — no
- * epsilon seams) plus a fused park detent at the pocket's mouth edge.
- */
-function cutSlideChannel(
-  scope: DisposalScope,
-  solid: Shape3D,
-  ctx: {
-    centerX: number;
-    centerY: number;
-    topZ: number;
-    mouth: { sign: 1 | -1; extendMm: number };
-  },
-  pocketW: number,
-  pocketD: number
-): Shape3D {
-  const r = LABEL_PLATE_CORNER_RADIUS_MM;
-  const lipT = LABEL_SOCKET_LIP_THICKNESS_MM;
-  const overhang = LABEL_SOCKET_LIP_OVERHANG_MM;
-  const cavityTop = ctx.topZ - lipT;
-  const floorZ = cavityTop - LABEL_SOCKET_SLIDE_Z_CLEARANCE_MM - LABEL_SOCKET_POCKET_DEPTH_MM;
-  const { sign, extendMm } = ctx.mouth;
-
-  // Cavity: pocket + mouth corridor at full plate width, up to the lip
-  // underside.
-  const cavityD = pocketD + extendMm;
-  const cavity = scope.register(
-    translate(
-      scope.register(
-        sketch(drawRoundedRectangle(pocketW, cavityD, r), 'XY', floorZ).extrude(cavityTop - floorZ)
-      ),
-      [ctx.centerX, ctx.centerY + (sign * extendMm) / 2, 0]
-    )
-  );
-  let result = scope.register(unwrap(cut(solid as ValidSolid, cavity as ValidSolid)));
-
-  // Lip window: inset by the overhang on the two side walls and the
-  // anchor-side wall; open through the mouth. Cut from the cavity top up
-  // past the shelf top.
-  const windowW = pocketW - 2 * overhang;
-  const windowD = pocketD - overhang + extendMm;
-  const window = scope.register(
-    translate(
-      scope.register(
-        sketch(drawRoundedRectangle(windowW, windowD, r), 'XY', cavityTop).extrude(lipT + 1)
-      ),
-      [ctx.centerX, ctx.centerY + (sign * (overhang + extendMm)) / 2, 0]
-    )
-  );
-  result = scope.register(unwrap(cut(result, window as ValidSolid)));
-
-  // Park detent: a low bar across the corridor floor just past the pocket
-  // edge — the plate rides over it on the way in and rests behind it.
-  const detentW = windowW;
-  const detentCenterY = ctx.centerY + sign * (pocketD / 2 + LABEL_SOCKET_DETENT_DEPTH_MM / 2);
-  const detent = scope.register(
-    translate(
-      scope.register(
-        sketch(
-          drawRoundedRectangle(detentW, LABEL_SOCKET_DETENT_DEPTH_MM, 0.2),
-          'XY',
-          floorZ
-        ).extrude(LABEL_SOCKET_DETENT_HEIGHT_MM)
-      ),
-      [ctx.centerX, detentCenterY, 0]
-    )
-  );
-  result = scope.register(unwrap(fuse(result, detent as ValidSolid)));
   return result;
 }
